batonez: can now create and feedforward a neural network
gen - functions to create neural nets and init them with weights and biases
dump - functions to dump neural net layer by layer, optionally with input and
       activation vectors
mymath - any math. Currently a couple of vector math funcs and a sigmoid activation
         function
neural - feedforward() method

A neural network is represented as a list of layers.
A layer is a dict with two keys: "weights" and "biases"
Under "weights" is a matrix of weights for this layer. Every
row corresponds to one neuron. Number of rows equals the number of neurons in the layer.
Under "biases" is a vector of biases for this layer. It's length must be equal
to number of neurons in the layer.

Vectors and matrices are presented as python lists. A matrix is stored as
a single list row-by-row. So if i is a row index and j is a col index, M is
the matrix and mat is it's list representation then to access an element M[i,j]
you should get mat[i*cols + j] element from the list, where cols is the number of
columns in the matrix (which equals to the number of inputs for every neuron in the layer
in our case).